// src/app/Enums/height.ts
export enum Height
{
    FourFeet = 0,

    FourFeet1Inch = 1,

    FourFeet2Inches = 2,

    FourFeet3Inches = 3,

    FourFeet4Inches = 4,

    FourFeet5Inches = 5,

    FourFeet6Inches = 6,

    FourFeet7Inches = 7,

    FourFeet8Inches = 8,

    FourFeet9Inches = 9,

    FourFeet10Inches = 10,

    FourFeet11Inches = 11,

    FiveFeet = 12,

    FiveFeet1Inch = 13,

    FiveFeet2Inches = 14,

    FiveFeet3Inches = 15,

    FiveFeet4Inches = 16,

    FiveFeet5Inches = 17,

    FiveFeet6Inches = 18,

    FiveFeet7Inches = 19,

    FiveFeet8Inches = 20,

    FiveFeet9Inches = 21,

    FiveFeet10Inches = 22,

    FiveFeet11Inches = 23,

    SixFeet0Inches = 24,

    SixFeet1Inches = 25,

    SixFeet2Inches = 26,

    SixFeet3Inches = 27,

    SixFeet4Inches = 28,

    SixFeet5Inches = 29,

    SixFeet6Inches = 30,
}